Phase 4: hook out sort/filters, ordering and browsing
Three more cohesive slices of ArcadeCatalog's state, each with its own
persistence, effect and derived values:

- useCatalogSortAndFilters  sort mode and filters (both written back to
  localStorage on change), the sort menu's open state and the
  outside-tap/Escape listener that closes it
- useCatalogOrdering        desiredOrder, userOrderKey, and the
  commit-once-per-key layout effect that paints displayedEntries; the
  two refs it needs stay private to the hook
- useCatalogBrowsing        the category filter, the browse page and its
  reset, the scroll helpers, and shelfCategories/browseEntries/
  pageEntries

RAIL_CARD_LIMIT is now exported from CatalogRail.tsx, the module that
owns rails, since useCatalogBrowsing slices shelves to it and the curated
rails still slice to it too.

catalogPending, layoutReady and showCurated deliberately stay in
ArcadeCatalog. They fold readiness from three different hooks — signals,
creator shelf, featured pool — and a hook around them would take all
three flags back as arguments and hand one boolean back, relocating the
coupling instead of removing it. Same call as the drafts/autosave machine
in CodeSurface (cb77556).

ArcadeCatalog 545 -> 418. No behaviour change: the browse-page reset
still runs after the order commit, since layout effects precede passive
ones however the hooks are ordered.
